Histórico de transações

+++ src/IConta.java
@@ -1,11 +1,12 @@
+import java.math.BigDecimal;
 
 public interface IConta {
 	
-	void sacar(double valor);
+	void sacar(BigDecimal valor);
 	
-	void depositar(double valor);
+	void depositar(BigDecimal valor);
 	
-	void transferir(double valor, IConta contaDestino);
+	void transferir(BigDecimal valor, IConta contaDestino);
 	
 	void imprimirExtrato();
 }
