+++ src/Main.java
@@ -1,3 +1,4 @@
+import java.math.BigDecimal;
 
 public class Main {
 
@@ -8,8 +9,8 @@ public class Main {
 		Conta cc = new ContaCorrente(venilton);
 		Conta poupanca = new ContaPoupanca(venilton);
 
-		cc.depositar(100);
-		cc.transferir(100, poupanca);
+		cc.depositar(BigDecimal.valueOf(100));
+		cc.transferir(BigDecimal.valueOf(100), poupanca);
 		
 		cc.imprimirExtrato();
 		poupanca.imprimirExtrato();
